Extract AttributeVersionRequirementHelper
