+++ extension.neon
@@ -80,6 +80,12 @@ services:
 	-
 		class: PHPStan\Type\PHPUnit\DynamicCallToAssertionIgnoreExtension
 
+	-
+		class: PHPStan\Rules\PHPUnit\AttributeVersionRequirementHelper
+		arguments:
+			deprecationRulesInstalled: %deprecationRulesInstalled%
+			bleedingEdge: %featureToggles.bleedingEdge%
+
 conditionalTags:
 	PHPStan\PhpDoc\PHPUnit\MockObjectTypeNodeResolverExtension:
 		phpstan.phpDoc.typeNodeResolverExtension: %phpunit.convertUnionToIntersectionType%
