+++ rules.neon
@@ -27,9 +27,6 @@ services:
 
 	-
 		class: PHPStan\Rules\PHPUnit\AttributeRequiresPhpVersionRule
-		arguments:
-			deprecationRulesInstalled: %deprecationRulesInstalled%
-			bleedingEdge: %featureToggles.bleedingEdge%
 		tags:
 			- phpstan.rules.rule
